Linked list is needed where we have a large number of data to store without its count and need to go into the next node or previous node easily. As addition and removal operations are faster in linked lists than arrays, these need to be applied first when we are talking about linked lists.
Linked list can be used to go into the next and previous pages in any application. Taking a simple example of the NEXT and PREVIOUS icons available on web pages that allow us to go on that section. Here linked list will be used to go into the next node (NEXT page) and previous node (PREVIOUS page).